Allow author to create source battery, and create a default one if they don't
Previously we were adding the battery on the fly when a measurement was made.
Now we allow the authors to add a component with the UID 'source' which we will
assume to be a/the power source, and we will connect it to the left rails by
default. If no component with the UID 'source' is detected, we will add a default
9V battery.

Tests are updated to reflect new component always in breadboard, and ability to add
authored or default power sources.