Add horizontal dragging to oscope view.
Dragging the mouse across the oscope screen will offset the
phase of the trace. A full swipe across the screen will move
the trace a full 2*PI.

This means that the actual distance the trace moves on the screen
will be different depending on the angle of the trace being displayed.
However, this feels more intuitive at most zoom levels.