Remove setting of logged-in user
Short- and medium-term future we are getting rid of the
SPARKS portal, so there is no need for logged-in users.
However, this commit does not remove all the associated
code, in case we ever want to restore this functionality,
but merely removes the setting of the user (e.g. for
users with cookies still set).