Refactoring: Initial conversion to CommonJS format and browserify build
This is a 90% conversion. Some modules, such as breadboard.js and all the
third-party libraries, still place objects in the global namespace.

However, the core MVC classes are now all shared using require();

There's a little ugliness with a circular dependency on workbenchController,
which causes us to have to require() it later in some classes, or it won't
have been initialized.

The project can be built using

browserify app/init.js --standalone sparks > sparks.js

(Built file in the next commit)