Initial refactoring of breadboard model into CommonJS module
This removes the global namespace pollution of the two main breadboard
APIs (breadModel and getBreadBoard), exporting them into an api that
can be required.

This also cleans up the breadboard model file.

We still need to separate out the model from the controller, and eliminate
the old Flash-style breadModel API.