Hack-ish fix for relative path loading issue
The JS needs to load resources, but scripts relative url references
are always relative to the html page. This is a problem if breadboard
is in some nested vendor folder.

This uses the DOM's script tag to find the path for the sparks.js
script.

It assumes that the /public folder is left intact.

Note that this won't work if sparks.js is loaded in some dynamic way