Updated multimeter to display negative voltages
Added tracking of actual value along with existing absolute value in
the multimeter component.  The absolute value is still used as is in
the code with the actual value providing the sign before the display
is updated.

A small change was also needed to allow both the HV and negative sign
to display at the same time.  If both need to display the second blank
space is used for the sign, eg 'h 3 0 0' becomes 'h-3 0 0'.