Fixed bug that caused two editor tooltips to show when the component was double clicked [#123226121]
This only manifested when the select option was added to the component editor.  With the slider both tooltips closed when the mouseleave event triggered but the select element sometimes caused a focus change which prevented the hidden tooltip from closing.  This made the UI look like it "reset" the edited field.